+++ src/main/webapp/pages/projects/project_new.html
@@ -1,16 +1,21 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      data-layout-decorate="~{template/page}">
-<head>
+<html
+  xmlns:th="http://www.thymeleaf.org"
+  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+  data-layout-decorate="~{template/page}"
+>
+  <head>
     <title th:text="#{projects.create.heading}">THIS IS SOMETHING WRONG</title>
-    <link rel="stylesheet" th:href="@{/dist/css/project-new.bundle.css}">
-</head>
-<body>
-<div id="root" style="height: 100%; background-color: darkgrey" layout:fragment="page">
-</div>
-<th:block layout:fragment="scripts" th:inline="javascript">
-    <script th:src="@{/dist/js/project-new.bundle.js}"></script>
-</th:block>
-</body>
+    <link rel="stylesheet" th:href="@{/dist/css/project-create.bundle.css}" />
+  </head>
+  <body>
+    <div
+      id="root"
+      style="height: 100%; min-height: 100%"
+      layout:fragment="page"
+    ></div>
+    <th:block layout:fragment="scripts" th:inline="javascript">
+      <script th:src="@{/dist/js/project-create.bundle.js}"></script>
+    </th:block>
+  </body>
 </html>
